Sales Track Create Operation

+++ database/migrations/2025_02_19_060542_create_sales_tracks_table.php
@@ -17,11 +17,11 @@ return new class extends Migration
             $table->foreignId('business_id')->constrained('businesses')->cascadeOnDelete();
             $table->foreignId('user_id')->constrained('users')->cascadeOnDelete();
             $table->foreignId('property_id')->constrained('properties')->cascadeOnDelete();
-            $table->float('price');
+            $table->decimal('price',5,2);
             $table->enum('status', ['active', 'pending', 'close', 'expired'])->default('active');
             $table->date('date_under_contract');
             $table->date('closing_date');
-            $table->float('purchase_price');
+            $table->decimal('purchase_price');
             $table->string('buyer_seller')->nullable();
             $table->decimal('referral_fee_pct', 5,2);
             $table->decimal('commission_on_sale', 5,2);
